refactor(translation): canonicalize streaming ownership
