+++ scripts/tests/test_refactor_baselines.py
@@ -44,6 +44,10 @@ class RefactorBaselineTests(unittest.TestCase):
             "src/modules/translation/include/aimee/translation/aimee_frontend.h",
             files,
         )
+        self.assertIn(
+            "src/modules/translation/include/aimee/translation/aimee_ir_stream.h",
+            files,
+        )
 
     def test_repository_baseline_is_cwd_independent(self) -> None:
         with tempfile.TemporaryDirectory() as cwd:
