+++ src/modules/translation/aimee_ir_stream.c
@@ -1,5 +1,5 @@
 /* aimee_ir_stream.c -- see aimee_ir_stream.h. */
-#include "aimee_ir_stream.h"
+#include <aimee/translation/aimee_ir_stream.h>
 
 #include <aimee/translation/aimee_backend.h> /* converse_stop_reason (shared with the non-stream parse) */
 #include "cJSON.h"
