+++ src/server/anthropic_http.c
@@ -35,7 +35,7 @@
 #include <aimee/ir/aimee_ir_metrics.h>
 #include "aimee_ir_serve.h"                  /* Slice 5: IR live request-build */
 #include <aimee/translation/aimee_backend.h> /* Slice 3: openai_backend_parse (IR response path) */
-#include "aimee_ir_stream.h"                 /* Slice 5-wire: IR-delta incremental relay */
+#include <aimee/translation/aimee_ir_stream.h> /* Slice 5-wire: IR-delta incremental relay */
 #include "ingress_preinject.h"
 #include "json_fluent.h"
 #include "log.h"
