+++ src/tests/test_aimee_converse_stream.c
@@ -7,7 +7,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "aimee_ir_stream.h"
+#include <aimee/translation/aimee_ir_stream.h>
 #include "cJSON.h"
 
 /* Parse `json` and run one ConverseStream event through the parser. Returns the
